+++ src/common/Back/Back.tsx
@@ -1,10 +1,12 @@
 import { useCallback } from 'react';
+import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Icon from 'semantic-ui-react/dist/commonjs/elements/Icon';
 import useStore from 'store';
 
 function Back(): JSX.Element {
+  const { t } = useTranslation('common');
   const navigate = useNavigate();
   const darkMode = useStore(useCallback((state) => state.darkMode, []));
 
@@ -15,7 +17,7 @@ function Back(): JSX.Element {
   return (
     <Button onClick={handleBack} style={{ height: 36 }} inverted={darkMode} type="button">
       <Icon name="arrow left" />
-      Back
+      {t('back')}
     </Button>
   );
 }